+++ lib/tests/integration.rs
@@ -1,6 +1,7 @@
 use kron_lib::*;
 
 const EPOCH_ISO8601: &str = "1970-01-01T00:00:00Z";
+const EPOCH_ISO8601NS: &str = "1970-01-01T00:00:00.000000000Z";
 
 #[test]
 fn epoch_display() {
@@ -19,6 +20,11 @@ fn epoch_timestamp() {
 #[test]
 fn epoch_format_iso8601() {
     assert_eq!(Kron::epoch().format(&ISO8601).unwrap(), EPOCH_ISO8601);
+}
+
+#[test]
+fn epoch_format_iso8601ns() {
+    assert_eq!(Kron::epoch().format(&ISO8601NS).unwrap(), EPOCH_ISO8601NS);
 }
 
 #[test]