+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 * 1.1.0 (2023-05-26): Enable accessing the internal `time::OffsetDateTime`; add
   `CHANGELOG.md`; fix readme
 * 2.0.0 (2023-05-28): Split to `kron-lib` and `kron` crates
+* 2.1.0 (2023-05-29): lib: Add ISO8601NS format, `KronFormat::from()`,
+  `Kron::from()`; cli: Add multiple args, `-f` and `-r` options
+